+++ app/components/flat_pack/comments/composer/component.rb
@@ -54,19 +54,19 @@ module FlatPack
         end
 
         def toolbar(*args, **kwargs, &block)
-          return get_slot(:toolbar) if args.empty? && kwargs.empty? && !block_given?
+          return get_slot(:toolbar) if args.empty? && kwargs.empty? && !block
 
           set_slot(:toolbar, nil, *args, **kwargs, &block)
         end
 
         def attachments(*args, **kwargs, &block)
-          return get_slot(:attachments) if args.empty? && kwargs.empty? && !block_given?
+          return get_slot(:attachments) if args.empty? && kwargs.empty? && !block
 
           set_slot(:attachments, nil, *args, **kwargs, &block)
         end
 
         def actions(*args, **kwargs, &block)
-          return get_slot(:actions) if args.empty? && kwargs.empty? && !block_given?
+          return get_slot(:actions) if args.empty? && kwargs.empty? && !block
 
           set_slot(:actions, nil, *args, **kwargs, &block)
         end